Moved JSR-305 related annotations to a separate package
