+++ context-propagation/src/main/java/io/micrometer/context/DefaultContextSnapshot.java
@@ -19,6 +19,8 @@ import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Predicate;
 
+import io.micrometer.context.util.annotation.Nullable;
+
 /**
  * Default implementation of {@link ContextSnapshot}.
  *
