+++ context-propagation/src/main/java/io/micrometer/context/DefaultContextSnapshotFactory.java
@@ -20,6 +20,8 @@ import java.util.Map;
 import java.util.Objects;
 import java.util.function.Predicate;
 
+import io.micrometer.context.util.annotation.Nullable;
+
 /**
  * Default implementation of {@link ContextSnapshotFactory}.
  *
