+++ context-propagation/src/main/java/io/micrometer/context/ThreadLocalAccessor.java
@@ -18,6 +18,8 @@ package io.micrometer.context;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
+import io.micrometer.context.util.annotation.Nullable;
+
 /**
  * Contract to assist with setting and clearing a {@link ThreadLocal}.
  *
