+++ context-propagation/src/main/java/io/micrometer/context/util/annotation/NonNullFields.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.micrometer.context;
+package io.micrometer.context.util.annotation;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
