+++ context-propagation/src/test/java/io/micrometer/context/TestContextAccessor.java
@@ -18,6 +18,8 @@ package io.micrometer.context;
 import java.util.Map;
 import java.util.function.Predicate;
 
+import io.micrometer.context.util.annotation.Nullable;
+
 /**
  * ThreadLocalAccessor for testing purposes with a given key and {@link ThreadLocal}
  * instance.
